Add display-transformer for mediator-open-with (#18)
I have just now written [a simple package](https://github.com/dalanicolai/mediator) that uses xdg mime type information to present a list of fitting applications for opening a file (on GNU/linux).
The functionality can be used as an Ivy or Embark (etc.) action. It also support icons in Ivy using this package, but for that, currently, the user has to add the right definition to the display-transformers list. It would be nice if this can be added to the list by default, hence this PR.

I also have a small question. The icons only show when invoking the `mediator-open-with` command directly, so not when invoked as an Embark or Ivy action (although for Embark I defined the `embark-open-with` function, which has no display-transformer definition). Do you know how to get the icons also when invoking the command as an Ivy action? Thank you...